test(django-cf): fold the worker harness into one conftest
tests/in_worker_harness.py and tests/in_worker/conftest.py were a second copy of
what tests/conftest.py already did: two ways to start `pywrangler dev`, two
readiness probes, two teardowns. They are now one file.

The two start-up modes are not interchangeable and both survive - the app
fixtures need uv project mode for their collectstatic build step, the in-worker
fixture needs --no-project plus hand-vendored libraries - but they now share
port allocation, the readiness poll and the process-group kill that keeps
workerd from being orphaned. `dev_server` becomes `in_worker_server` and points
at tests/in_worker/worker directly, which retires the `worker_project_dir`
indirection.

Python 3.12 leaves the in-worker matrix. That runtime (Pyodide 0.26.0a2) has no
JSPI, so pyodide.ffi.run_sync is missing and all 15 R2Storage tests skip
themselves; the rest duplicates the 3.13 and 3.14 runs. 505 collected tests
become 348 and tests/in_worker drops from ~50s to ~33s. This is not the
runtime-sdk's reason for excluding 3.12: these suites contain no async tests,
so there were no false passes to fix here.

The npm setup-* scripts go too. The harness copies django_cf into a tmpdir per
run, so nothing has to be staged into templates/ or tests/servers/ first, and
AGENTS.md no longer documents a step that does not exist.

346 passed, 2 skipped. Both skips are the strict xfail in
tests/in_worker/worker/src/test_db.py, one per remaining config.
