+++ packages/django-cf/package.json
@@ -12,11 +12,7 @@
   },
   "license": "MIT",
   "scripts": {
-    "setup-durable-objects": "cd templates/durable-objects && npm run dependencies && rm -rf python_modules/django_cf && cp -r ../../django_cf python_modules/django_cf",
-    "setup-d1": "cd templates/d1 && npm run dependencies && rm -rf python_modules/django_cf && cp -r ../../django_cf python_modules/django_cf",
-    "setup-test-servers": "cd tests/servers/r2 && npm run dependencies && rm -rf python_modules/django_cf && cp -r ../../../django_cf python_modules/django_cf",
-    "setup-test": "uv sync && npm run setup-durable-objects && npm run setup-d1 && npm run setup-test-servers",
-    "test": "pytest",
+    "test": "uv run pytest",
     "upgrade-templates": "cd templates/durable-objects && uv add django-cf --upgrade && cd ../d1 && uv add django-cf --upgrade",
     "build": "rm -rf dist/ && python3 -m build",
     "publish": "python3 -m twine upload dist/*"
