--- packages/django-cf/tests/in_worker/conftest.py
@@ -1,13 +1,0 @@
-from tests.in_worker_harness import (
-    CompatConfig,
-    compat_config,
-    dev_server,
-    worker_project_dir,
-)
-
-__all__ = [
-    "CompatConfig",
-    "compat_config",
-    "dev_server",
-    "worker_project_dir",
-]
